reorganize composition
